added font, organized component directory structure
