+++ App.js
@@ -1,10 +1,15 @@
 import React from 'react';
-import {StyleSheet, View} from 'react-native';
-import {Text, Icon} from 'react-native-elements';
-import TodoList from './src/components/TodoList';
+import {StyleSheet, View, Text} from 'react-native';
+import {Icon} from 'react-native-elements';
+import TodoList from './src/components/TodoList/TodoList';
+import { Font } from 'expo';
+import TodayText from './src/components/TodayText/TodayText'
+
 
 export default class App extends React.Component {
+
     state = {
+        fontLoaded: false,
         list: [
             {
                 title: 'Take out the trash',
@@ -20,6 +25,14 @@ export default class App extends React.Component {
             }
         ]
     };
+    async componentDidMount() {
+        await Font.loadAsync({
+            'WaitingForTheSunrise': require('./src/assets/fonts/WaitingForTheSunrise.ttf'),
+        });
+
+        this.setState({ fontLoaded: true });
+    }
+
 
     handleAddItem = (item) => {
         console.log(item)
@@ -33,16 +46,25 @@ export default class App extends React.Component {
         });
     };
 
+
     render() {
         return (
-            <View style={styles.container}>
-                <Text h2>Today!</Text>
-                <TodoList items={this.state.list} onCheckItem={this.handleCheckItem}/>
-                <Icon
-                    name='add-circle'
-                    color='#f50'
-                    onPress={() => this.handleAddItem('TODO: pop a modal to capture new item.  also, style this button plz')}/>
-            </View>
+
+            <React.Fragment>
+                {this.state.fontLoaded ? (
+                    <View style={styles.container}>
+                        <TodayText h2 style={{fontSize: 56}}>Today!</TodayText>
+
+
+                        <TodoList items={this.state.list} onCheckItem={this.handleCheckItem}/>
+                        <Icon
+                            name='add-circle'
+                            color='#f50'
+                            onPress={() => this.handleAddItem('TODO: pop a modal to capture new item.  also, style this button plz')}/>
+                    </View>
+                    ) : null
+                }
+            </React.Fragment>
         );
     }
 }
@@ -53,7 +75,11 @@ const styles = StyleSheet.create({
         backgroundColor: '#fff',
         alignItems: 'center',
         justifyContent: 'flex-start',
-        marginTop: 40
+        marginTop: 40,
+    },
+    headline: {
+        fontSize:62,
+        color:'red'
     },
     list: {
         width: 200
